% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{An integer or numeric scalar}

\item{b}{An integer or numeric scalar}
}
\value{
The greatest common divisor of a and b as a numeric value
}
\description{
Finds the greatest common divisor of two integers using the Euclidean algorithm.
The algorithm repeatedly applies the division algorithm until the remainder is zero.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
