% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{A test graph.}
\format{
A data frame with 3 columns.
\describe{
  \item{v1}{vertex 1}
  \item{v2}{vertex 2}
  \item{w}{edge weight}
  ...
}
}
\source{
\url{https://raw.githubusercontent.com/STIMALiU/AdvRCourse/master/2020_LecturesLabs/732A94_AdvancedRHT2020_Lab03.pdf}
}
\usage{
wiki_graph
}
\description{
A test graph.
}
\keyword{datasets}
