% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{\code{dijkstra} returns the distance between two nodes for a given path.}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{Graph.}

\item{init_node}{Starting node.}
}
\value{
Array indicating path length to each node.
}
\description{
The pseudocode used for this function can be found on 
\url{https://en.wikipedia.org/wiki/Dijkstra's_algorithm}
}
\examples{
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)
}
