% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{\code{euclidean} returns the lowest common denominator between two numbers.}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{B number.}
}
\value{
The lowest common denominator of \code{a} and \code{b}.
}
\description{
The pseudocode for this function can be found on the following link:
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\examples{
euclidean(1, 1)
euclidean(10, 1)
}
