% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{\code{euclidean} returns the lowest common denominator between two numbers.}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
The lowest common denominator of \code{x} and \code{y}.
}
\description{
\code{euclidean} returns the lowest common denominator between two numbers.
}
\examples{
euclidean(1, 1)
euclidean(10, 1)
}
